# backend/openapi.yaml
openapi: 3.1.0
info:
  title: NoesisForge API
  description: |
    NoesisForge backend API documentation. This API provides authentication, user management, and role-based access control functionality.
    
    ## Authentication
    This API uses JWT (JSON Web Token) for authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to 100 requests per minute per IP address.
    
  version: 1.0.0
  contact:
    name: NoesisForge API Support
    email: support@noesisforge.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.noesisforge.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique role identifier
        name:
          type: string
          description: Role name (system identifier)
        displayName:
          type: string
          description: Human-readable role name
        description:
          type: string
          description: Role description
        isDefault:
          type: boolean
          description: Is this the default role for new users
        isSystem:
          type: boolean
          description: Is this a system role (cannot be deleted)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique permission identifier
        name:
          type: string
          description: Permission name (system identifier)
        displayName:
          type: string
          description: Human-readable permission name
        description:
          type: string
          description: Permission description
        category:
          type: string
          description: Permission category
        isSystem:
          type: boolean
          description: Is this a system permission
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Detailed error information (optional)
...
    CreateRoleRequest:
      type: object
      required:
        - name
        - displayName
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
          description: Role name (system identifier)
        displayName:
          type: string
          minLength: 2
          maxLength: 100
          description: Human-readable role name
        description:
          type: string
          description: Role description
        permissionIDs:
          type: array
          items:
            type: string
            format: uuid
          description: List of permission IDs to assign
    UpdateRoleRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 100
          description: Human-readable role name
        description:
          type: string
          description: Role description
        permissionIDs:
          type: array
          items:
            type: string
            format: uuid
          description: List of permission IDs to assign
    AssignRoleRequest:
      type: object
      required:
        - userID
        - roleID
      properties:
        userID:
          type: string
          format: uuid
          description: User ID
        roleID:
          type: string
          format: uuid
          description: Role ID
paths:
...
  /roles:
    get:
      summary: Get all roles
      description: Returns all roles with their permissions (Admin only)
      tags:
        - Roles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Roles retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create new role
      description: Creates a new role with specified permissions (Admin only)
      tags:
        - Roles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
            example:
              name: moderator
              displayName: Moderator
              description: Can moderate content and manage users
              permissionIDs:
                - 550e8400-e29b-41d4-a716-446655440001
                - 550e8400-e29b-41d4-a716-446655440002
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Role created successfully
                  role:
                    $ref: '#/components/schemas/Role'
        '400':
          description: Validation error or creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Role name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /roles/{id}:
    get:
      summary: Get role by ID
      description: Returns a specific role with its permissions (Admin only)
      tags:
        - Roles
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Role retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  role:
                    $ref: '#/components/schemas/Role'
        '400':
          description: Invalid role ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update role
      description: Updates a role's information and permissions (Admin only)
      tags:
        - Roles
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
            example:
              displayName: Updated Moderator
              description: Updated description
              permissionIDs:
                - 550e8400-e29b-41d4-a716-446655440001
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Role updated successfully
                  role:
                    $ref: '#/components/schemas/Role'
        '400':
          description: Validation error or update failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required or cannot modify system role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete role
      description: Deletes a role (Admin only, cannot delete system roles or roles with assigned users)
      tags:
        - Roles
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Role deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Role deleted successfully
        '400':
          description: Invalid role ID format or deletion failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required or cannot delete system role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /roles/permissions:
    get:
      summary: Get all permissions
      description: Returns all available permissions (Admin only)
      tags:
        - Roles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Permissions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  permissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Permission'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /roles/permissions/categories/{category}:
    get:
      summary: Get permissions by category
      description: Returns permissions filtered by category (Admin only)
      tags:
        - Roles
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: path
          required: true
          description: Permission category
          schema:
            type: string
            enum: [document, search, chat, admin]
      responses:
        '200':
          description: Permissions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  permissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Permission'
        '400':
          description: Category is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /roles/assign:
    post:
      summary: Assign role to user
      description: Assigns a role to a specific user (Admin only)
      tags:
        - Roles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
            example:
              userID: 550e8400-e29b-41d4-a716-446655440000
              roleID: 550e8400-e29b-41d4-a716-446655440001
      responses:
        '200':
          description: Role assigned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Role assigned successfully
        '400':
          description: Validation error or assignment failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User or role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
security:
  - BearerAuth: [] 